More craterpy refactoring
